Multiple refactory
- fix camera projection
- change pivot point to center (rotation center and render center)
- add enum dropdown support for ImGui
- fix ImGui transform rotation
- ctrl + D to duplicate game object
- DEL to delete game object
- update z-index (and rendering) at runtime

+++ src/main/java/components/GridLines.java
@@ -15,20 +15,20 @@ public class GridLines extends Component {
         Vector2f cameraPos = camera.position;
         Vector2f projectionSize = camera.getProjectionSize();
 
-        int firstX = (int) cameraPos.x / Settings.GRID_WIDTH * Settings.GRID_WIDTH - Settings.GRID_WIDTH;
-        int firstY = (int) cameraPos.y / Settings.GRID_HEIGHT * Settings.GRID_HEIGHT - Settings.GRID_HEIGHT;
+        float firstX = (int) cameraPos.x / Settings.GRID_WIDTH * Settings.GRID_WIDTH - Settings.GRID_WIDTH;
+        float firstY = (int) cameraPos.y / Settings.GRID_HEIGHT * Settings.GRID_HEIGHT - Settings.GRID_HEIGHT;
 
-        int numVerticalLines = (int) (projectionSize.x * camera.getZoom()) / Settings.GRID_WIDTH + 2;
-        int numHorizontalLines = (int) (projectionSize.y * camera.getZoom()) / Settings.GRID_HEIGHT + 2;
+        int numVerticalLines = (int) (projectionSize.x * camera.getZoom() / Settings.GRID_WIDTH) + 2;
+        int numHorizontalLines = (int) (projectionSize.y * camera.getZoom() / Settings.GRID_HEIGHT) + 2;
 
-        int height = (int) (projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT * 2;
-        int width = (int) (projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH * 2;
+        float height = (int) (projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT * 2;
+        float width = (int) (projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH * 2;
 
         int maxNumLines = Math.max(numHorizontalLines, numVerticalLines);
         Vector3f color = new Vector3f(0.2f, 0.2f, 0.2f);
         for (int i = 0; i < maxNumLines; i++) {
-            int x = firstX + Settings.GRID_WIDTH * i;
-            int y = firstY + Settings.GRID_HEIGHT * i;
+            float x = firstX + Settings.GRID_WIDTH * i;
+            float y = firstY + Settings.GRID_HEIGHT * i;
 
             if (i < numVerticalLines) {
                 DebugDraw.addLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), color);
